requirements.txt
setuptools==70.0.0
faster-whisper==1.0.3
librosa==0.10.1
numpy==1.26.4
pydantic==1.10.15
click==8.1.7
typer==0.12.3
rich==13.7.1
soundfile==0.12.1
pandas==2.2.2
pyarrow==16.1.0
scipy==1.11.4
PyYAML==6.0.1
pytest==8.3.3
